Different fixes and changes

+++ remove.php
@@ -1,10 +1,7 @@
 <?php
-$path = $_GET['path'];
-if(is_file($path)){
-	unlink($path);
-}
-else if(is_dir($path)){
-	rmdir($path);
-}
+if (isset($_GET['path'])) {
+	$path = $_GET['path'];
 
-?>
+	if (is_file($path)) unlink($path);
+	else if (is_dir($path)) rmdir($path);
+}
